requirements.txt
datasets==2.11.0
torch==2.0.0
transformers==4.27.4
hydra-core==1.3.2
